requirements.txt
Flask==2.1.4
Flask-PyMongo==2.3.0
pymongo==3.12.0
gunicorn==19.7.1
PyJWT==1.7.1
